requirements.txt
asgiref==3.7.2
black==23.9.1
click==8.1.7
Django==4.2.5
django-tenants==3.5.0
djangorestframework==3.14.0
environs==9.5.0
marshmallow==3.20.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.8
pyscopg2==66.0.2
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
